// frontend/src/types/FuguesResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FugueDefinition } from "./FugueDefinition";
import type { FugueInfo } from "./FugueInfo";

/**
 * Response containing fugue state
 */
export type FuguesResponse = { infos: Array<FugueInfo>, definitions: Array<FugueDefinition>, };